pyproject.toml
[project]
name = "s3od"
version = "0.1.0"
description = "Simple and efficient background removal using S3OD"
readme = "README.md"
authors = [{ name = "S3OD Team" }]
requires-python = ">=3.8"
dependencies = [
  "huggingface-hub>=0.16.0",
  "numpy>=1.20.0",
  "opencv-python>=4.5.0",
  "pillow>=9.0.0",
  "torch>=2.0.0",
  "torchvision>=0.15.0",
]

[build-system]
requires = ["uv_build>=0.8.22,<0.9.0"]
build-backend = "uv_build"

[tool.uv.workspace]
members = ["synth_sod"]

[tool.uv.sources]
synth-sod = { workspace = true }

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "synth-sod",
]
